Separate read and write tests into dedicated files
