+++ spec/active_record/monetizable_spec.rb
@@ -803,193 +803,5 @@ if defined? ActiveRecord
         end
       end
     end
-
-    describe "#read_monetized" do
-      it "returns monetized attribute's value" do
-        reduced_price = product.read_monetized(:reduced_price, :reduced_price_cents)
-
-        expect(reduced_price).to be_an_instance_of(Money)
-        expect(reduced_price).to eq(Money.new(product.reduced_price_cents, product.reduced_price_currency))
-      end
-
-      context "when memoized" do
-        it "memoizes monetized attribute's value" do
-          product.instance_variable_set :@reduced_price, nil
-          reduced_price = product.read_monetized(:reduced_price, :reduced_price_cents)
-
-          expect(product.instance_variable_get(:@reduced_price)).to eq(reduced_price)
-        end
-
-        it "resets memoized attribute's value if amount has changed" do
-          reduced_price = product.read_monetized(:reduced_price, :reduced_price_cents)
-          product.reduced_price_cents = 100
-
-          expect(product.read_monetized(:reduced_price, :reduced_price_cents)).not_to eq(reduced_price)
-        end
-
-        it "resets memoized attribute's value if currency has changed" do
-          reduced_price = product.read_monetized(:reduced_price, :reduced_price_cents)
-          product.reduced_price_currency = "CAD"
-
-          expect(product.read_monetized(:reduced_price, :reduced_price_cents)).not_to eq(reduced_price)
-        end
-      end
-
-      context "with preserve_user_input set" do
-        around do |example|
-          MoneyRails::Configuration.preserve_user_input = true
-          example.run
-          MoneyRails::Configuration.preserve_user_input = false
-        end
-
-        it "has no effect if validation passes" do
-          product.price = "14"
-
-          expect(product.save).to be_truthy
-          expect(product.read_monetized(:price, :price_cents).to_s).to eq("14.00")
-        end
-
-        it "preserves user input if validation fails" do
-          product.price = "14,0"
-
-          expect(product.save).to be_falsy
-          expect(product.read_monetized(:price, :price_cents).to_s).to eq("14,0")
-        end
-      end
-
-      context "with a monetized attribute that is nil" do
-        let(:service) do
-          Service.create(discount_cents: nil)
-        end
-
-        # rubocop:disable RSpec/VerifiedDoubles
-        let(:default_currency_lambda) do
-          double :default_fallback, read_currency: nil
-        end
-        # rubocop:enable RSpec/VerifiedDoubles
-
-        let(:read_monetized) { service.read_monetized(:discount, :discount_cents, options) }
-
-        around do |example|
-          service # Instantiate instance which relies on Money.default_currency
-          original_default_currency = Money.default_currency
-          Money.default_currency = -> { default_currency_lambda.read_currency }
-          example.run
-          Money.default_currency = original_default_currency
-        end
-
-        context "when allow_nil options is set" do
-          let(:options) { { allow_nil: true } }
-
-          before do
-            allow(default_currency_lambda).to receive(:read_currency)
-          end
-
-          it "does not attempt to read the fallback default currency" do
-            read_monetized
-
-            expect(default_currency_lambda).not_to have_received(:read_currency)
-          end
-        end
-      end
-    end
-
-    describe "#write_monetized" do
-      let(:value) { Money.new(1_000, "LVL") }
-
-      it "sets monetized attribute's value to Money object" do
-        product.write_monetized :price, :price_cents, value, false, nil, {}
-
-        expect(product.price).to be_an_instance_of(Money)
-        expect(product.price_cents).to eq(value.cents)
-        # Because :price does not have a column for currency
-        expect(product.price.currency).to eq(Product.currency)
-      end
-
-      it "sets monetized attribute's value from a given Fixnum" do
-        product.write_monetized :price, :price_cents, 10, false, nil, {}
-
-        expect(product.price).to be_an_instance_of(Money)
-        expect(product.price_cents).to eq(1000)
-      end
-
-      it "sets monetized attribute's value from a given Float" do
-        product.write_monetized :price, :price_cents, 10.5, false, nil, {}
-
-        expect(product.price).to be_an_instance_of(Money)
-        expect(product.price_cents).to eq(1050)
-      end
-
-      it "resets monetized attribute when given blank input" do
-        product.write_monetized :price, :price_cents, nil, false, nil, { allow_nil: true }
-
-        expect(product.price).to be_nil
-      end
-
-      it "sets monetized attribute to 0 when given a blank value" do
-        currency = product.price.currency
-        product.write_monetized :price, :price_cents, nil, false, nil, {}
-
-        expect(product.price.amount).to eq(0)
-        expect(product.price.currency).to eq(currency)
-      end
-
-      it "does not memoize monetized attribute's value if currency is read-only" do
-        product.write_monetized :price, :price_cents, value, false, nil, {}
-
-        price = product.instance_variable_get(:@price)
-
-        expect(price).to be_an_instance_of(Money)
-        expect(price.amount).not_to eq(value.amount)
-      end
-
-      context "without a default currency" do
-        let(:product) { OtherProduct.new }
-
-        around do |example|
-          default_currency = Money.default_currency
-          Money.default_currency = nil
-
-          example.run
-
-          Money.default_currency = default_currency
-        end
-
-        it "errors a NoCurrency Error" do
-          expect do
-            product.write_monetized :price, :price_cents, 10.5, false, nil, {}
-          end.to raise_error(Money::Currency::NoCurrency, "must provide a currency")
-        end
-      end
-
-      describe "instance_currency_name" do
-        it "updates instance_currency_name attribute" do
-          product.write_monetized :sale_price, :sale_price_amount, value, false, :sale_price_currency_code, {}
-
-          expect(product.sale_price).to eq(value)
-          expect(product.sale_price_currency_code).to eq("LVL")
-        end
-
-        it "memoizes monetized attribute's value with currency" do
-          product.write_monetized :sale_price, :sale_price_amount, value, false, :sale_price_currency_code, {}
-
-          expect(product.instance_variable_get(:@sale_price)).to eq(value)
-        end
-
-        it "ignores empty instance_currency_name" do
-          product.write_monetized :sale_price, :sale_price_amount, value, false, "", {}
-
-          expect(product.sale_price.amount).to eq(value.amount)
-          expect(product.sale_price.currency).to eq(Product.currency)
-        end
-
-        it "ignores instance_currency_name that model does not respond to" do
-          product.write_monetized :sale_price, :sale_price_amount, value, false, :non_existing_currency, {}
-
-          expect(product.sale_price.amount).to eq(value.amount)
-          expect(product.sale_price.currency).to eq(Product.currency)
-        end
-      end
-    end
   end
 end
